Renamed parse.py for clarity in imports. Added plot.py to start creating a class for summary plots.
